requirements.txt
asgiref==3.8.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
mongoengine==0.27.0
mysql-connector-python==8.2.0
packaging==24.0
protobuf==4.21.12
PyJWT==2.8.0
pymongo==4.6.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
